Now purchases module partially working

+++ app/src/main/java/com/go/kchin/model/database/PurchaseOperation.java
@@ -9,7 +9,7 @@ import com.orm.SugarRecord;
 public class PurchaseOperation extends SugarRecord {
 
     public String purchaseConcept;
-    public int purchaseItems;
+    public float purchaseItems;
     public float purchaseAmount;
     public long purchaseDateTime;
